stop searches while update task runs, add option 'nosync' to skip rsync

+++ lib/tasks/scripts/update_freecen2_production.sh
@@ -1,9 +1,14 @@
 #!/usr/bin/env /usr/local/bin/bash
 # note: on ubuntu dev machine, needed to set up a link to /bin/bash
 
-# This script updates the freecen2 production database
-# It just rsyncs the data from FC1 to FC2 on the local machine before
-# running the rake task that updates the database.
+# This script updates the freecen2 production database.
+# It rsyncs the data from FC1 to FC2 on the local machine (unless 'nosync' is
+# passed in as argument 1 or 2 on the command line) before
+# running the rake tasks that:
+#  1) update the database
+#  2) initialize geolocation lat/long for new subplaces that aren't set
+#  3) update the places cache
+#
 #
 # To manually rsync the data from freecen production server to test2 server:
 #  ssh -A user@test2host
@@ -21,13 +26,6 @@ trace() {
   echo "[update_freecen2_production.sh][${HN}] ${NOW} $@" >&2
 }
 
-#fail() {
-#  sudo /root/bin/searchctl.sh enable
-#  trace "FATAL $@"
-#  exit 1
-#}
-#trap fail ERR
-
 FC1_DATA=/raid/freecen
 FC1_STAT_FILE=/home/apache/hosts/freecen/status/db-stats #fc1_coverage_stats value in config/mongo_config.yml should match
 FC2_DATA=/raid/freecen2
@@ -38,7 +36,7 @@ GEOLOCATION_FILE=${APP_ROOT}/test_data/Place_and_church_name_resources/places_fr
 WEB_USER=webserv
 BUNDLE=bundle
 #different directories on development machine (pass in "development" as arg 1)
-if [ $# -ge 1 ] && [ $1 == "development" ]; then
+if [ $# -ge 1 ] && [ "$1" == "development" ]; then
   trace "***NOTICE: using local development machine directory structure"
   FC1_DATA=~/freeUKGEN/data/update_test_fc1
   FC1_STAT_FILE=/home/apache/hosts/freecen/status/db-stats #ok if doesn't exist, just won't update the graphs without it.  Should match fc1_coverage_stats value in config/mongo_config.yml
@@ -74,26 +72,47 @@ if [[ ! -d ${LOG_DIR} ]] ; then
   mkdir -p ${LOG_DIR}
 fi
 
-# rsync the FC2 data from FC1 data directories
-trace "doing rsync of FreeCen1 metadata (ctyPARMS.DAT) files into FreeCen2"
-sudo -u ${WEB_USER} rsync -avz --delete ${FC1_DATA}/fixed ${FC2_DATA}/freecen1/ 2>${LOG_DIR}/rsync.errors | egrep -v '(^receiving|^sending|^sent|^total|^cannot|^deleting|^$|/$)' > ${LOG_DIR}/freecen1.delta
-
-trace "doing rsync of FreeCen1 validated piece (.VLD) files into FreeCen2"
-sudo -u ${WEB_USER} rsync -avz --delete ${FC1_DATA}/pieces ${FC2_DATA}/freecen1/ 2>${LOG_DIR}/rsync.errors | egrep -v '(^receiving|^sending|^sent|^total|^cannot|^deleting|^$|/$)' >> ${LOG_DIR}/freecen1.delta
-
-if [[ -f ${FC1_STAT_FILE} ]] ; then
-  trace "doing rsync of FreeCen1 db-status file into FreeCen2"
-  sudo -u ${WEB_USER} rsync -avz ${FC1_STAT_FILE} ${FC2_DATA}/freecen1/ 2>${LOG_DIR}/rsync.errors | egrep -v '(^receiving|^sending|^sent|^total|^cannot|^deleting|^$|/$)' >> ${LOG_DIR}/freecen1.delta
-else
-  trace "***WARNING: not doing rsync of status file because ${FC1_STAT_FILE} not found"
+DO_RSYNC=true
+if [ $# -ge 1 ] && [ "$1" == "nosync" ]; then
+  DO_RSYNC=false
+elif [ $# -ge 2 ] && [ "$2" == "nosync" ]; then
+  DO_RSYNC=false
 fi
 
-#Do we need to disable/enable searches below using /root/bin/searchctl.sh?
-#It was in the FreeReg2 script, but I don't think we need it. If we do, then
-#uncomment the lines and also uncomment the fail()/trap above.
+# rsync the FC2 data from FC1 data directories
+if [ "$DO_RSYNC" == true ] ; then
+  trace "doing rsync of FreeCen1 metadata (ctyPARMS.DAT) files into FreeCen2"
+  sudo -u ${WEB_USER} rsync -avz --delete ${FC1_DATA}/fixed ${FC2_DATA}/freecen1/ 2>${LOG_DIR}/rsync.errors | egrep -v '(^receiving|^sending|^sent|^total|^cannot|^deleting|^$|/$)' > ${LOG_DIR}/freecen1.delta
 
-#trace "disable of searches"
-#sudo /root/bin/searchctl.sh disable
+  trace "doing rsync of FreeCen1 validated piece (.VLD) files into FreeCen2"
+  sudo -u ${WEB_USER} rsync -avz --delete ${FC1_DATA}/pieces ${FC2_DATA}/freecen1/ 2>${LOG_DIR}/rsync.errors | egrep -v '(^receiving|^sending|^sent|^total|^cannot|^deleting|^$|/$)' >> ${LOG_DIR}/freecen1.delta
+
+  if [[ -f ${FC1_STAT_FILE} ]] ; then
+    trace "doing rsync of FreeCen1 db-status file into FreeCen2"
+    sudo -u ${WEB_USER} rsync -avz ${FC1_STAT_FILE} ${FC2_DATA}/freecen1/ 2>${LOG_DIR}/rsync.errors | egrep -v '(^receiving|^sending|^sent|^total|^cannot|^deleting|^$|/$)' >> ${LOG_DIR}/freecen1.delta
+  else
+    trace "***WARNING: not doing rsync of status file because ${FC1_STAT_FILE} not found"
+  fi
+
+else
+  trace "***WARNING: not doing rsync because 'nosync' option was passed into script"
+fi
+
+
+fail() {
+  sudo /root/bin/searchctl.sh enable
+  trace "FATAL $@"
+  exit 1
+}
+
+if [[ -f /root/bin/searchctl.sh ]] ; then
+  trap fail ERR
+  trace "disable searches while update script runs"
+  sudo /root/bin/searchctl.sh disable
+else
+  trace "***WARNING: /root/bin/searchctl.sh not found! not disabling searches"
+fi
+
 trace "running rake task to update the freecen database"
 sudo -u ${WEB_USER} ${BUNDLE} exec rake RAILS_ENV=production freecen_update_from_FC1["${FC2_DATA}/freecen1/fixed","${FC2_DATA}/freecen1/pieces"] --trace
 
@@ -106,7 +125,10 @@ sudo -u ${WEB_USER} ${BUNDLE} exec rake RAILS_ENV=production initialize_places_g
 trace "running rake task to update the places cache"
 sudo -u ${WEB_USER} ${BUNDLE} exec rake RAILS_ENV=production foo:refresh_places_cache["false"] --trace
 
-#trace "re enable searches"
-#sudo /root/bin/searchctl.sh enable
+if [[ -f /root/bin/searchctl.sh ]] ; then
+  trace "re-enable searches"
+  sudo /root/bin/searchctl.sh enable
+fi
+
 trace "finished"
 exit
